fix: GapDetector mapToDouble + ScannerService timestamp Long.parseLong
GapDetector: OHLCV.getVolume() returns double — use mapToDouble not mapToLong
ScannerService: OHLCV.getTimestamp() returns String — parse with Long.parseLong
before passing to Instant.ofEpochMilli()

+++ src/main/java/com/smcscanner/strategy/GapDetector.java
@@ -64,7 +64,7 @@ public class GapDetector {
         if (Math.abs(gapPct) < 0.3) return null;
 
         // Volume confirmation: compare first bar volume to average
-        double avgVol = prevBars.stream().mapToLong(OHLCV::getVolume).average().orElse(0);
+        double avgVol = prevBars.stream().mapToDouble(OHLCV::getVolume).average().orElse(0);
         double firstBarVol = todayBars.get(0).getVolume();
         double volRatio = avgVol > 0 ? firstBarVol / avgVol : 1.0;
